scan-frontend: factor out loading flow and test print flow
